feat(settings): add "Preload Effects" toggle
Add a setting which allows sequencer effect preload to be turned on/off
(on by default to match the current functionality).

In addition, refactor settings and hooks to a consistent format and
clean up the "main" script.

Closes #31
