+++ lang/en.json
@@ -1,5 +1,7 @@
 {
 	"lancer-weapon-fx.Sound Volume": "Sound Volume",
 	"lancer-weapon-fx.Sound Volume Hint": "Set a base volume level for all sound effects. Note that for each user, this value is combined with the user's \"Global: Interface\" volume setting (see the \"Playlists\" sidebar tab) to produce their final volume level.",
+	"lancer-weapon-fx.Preload Effects": "Preload Effects",
+	"lancer-weapon-fx.Preload Effects Hint": "If enabled, common effects (images and videos) will be pre-loaded when loading the world. Disabling this improves world loading time at the cost of lag the first time an animation is played.",
 	"lancer-weapon-fx.Debug: Play Miss Animations by Default": "Debug: Play Miss Animations by Default"
 }
